Huge memry optimization.

+++ VisualisationBinomialHeap/Models/BlockDataUtils.cs
@@ -2,7 +2,7 @@
 
 namespace MyMinecraft.Models;
 
-public enum BlockType : byte {
+public enum BlockType : ushort {
     AIR,
     DIRT,
     STONE,
